# inblog-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Inblog API",
    "version": "1.0.0",
    "description": "Manage blog content using the Inblog API.\n\n## 🚀 Getting Started\n\n### 1. Get API Key\nGet your API key from the dashboard:\n1. Login to [Inblog Dashboard](https://inblog.com/dashboard)\n2. Navigate to Settings → API Keys\n3. Click \"Create New API Key\" button\n4. Select required permission scopes\n5. Store your API key securely\n\n### 2. First Request\n```bash\ncurl -X GET https://api.inblog.ai/api/v1/me \\\n  -H \"Authorization: Bearer YOUR_API_KEY\"\n```\n\n### 3. Content-Type\n- **Request**: `Content-Type: application/json` or `application/vnd.api+json`\n- **Response**: `Content-Type: application/vnd.api+json`\n\n## 📋 Key Features\n\n- **Post Management**: Create, update, delete, and publish posts\n- **Tags and Authors**: Manage categories and author information\n- **Blog Settings**: Modify basic settings\n- **URL Redirects**: Configure 307 (temporary) and 308 (permanent) redirects\n- **JSON:API Standard**: Consistent response format\n\n## 🔐 Authentication\n\nAll API requests require Bearer token authentication via Authorization header:\n\n```\nAuthorization: Bearer YOUR_API_KEY\n```\n## 📚 Learn More\n\n- [JSON:API Specification](https://jsonapi.org/)",
    "contact": {
      "name": "Inblog Support",
      "email": "hello@inblog.ai"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://inblog.ai/api/v1",
...
    },
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/tags": {
      "get": {
        "tags": [
          "Tags"
        ],
        "summary": "List Tags",
        "description": "Retrieve all tags registered in the blog.",
        "operationId": "listTags",
        "security": [
          {
            "bearerAuth": [
              "tags:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Tag list retrieved successfully",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/TagListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Tags"
        ],
        "summary": "Create Tag",
        "description": "Create a new tag.\n\n**Required Fields:**\n- name: Tag name\n- slug: URL slug (must be unique within the blog)\n\n**Response Codes:**\n- 201: Tag created successfully\n- 409: Slug already exists\n",
        "operationId": "createTag",
        "security": [
          {
            "bearerAuth": [
              "tags:write"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreateRequest"
              }
            },
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tag created successfully",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowedError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/tags/{id}": {
      "get": {
        "tags": [
          "Tags"
        ],
        "summary": "Get Tag",
        "description": "Retrieve detailed information for a specific tag by ID.",
        "operationId": "getTag",
        "security": [
          {
            "bearerAuth": [
              "tags:read"
            ]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Tag ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tag retrieved successfully",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "tags": [
          "Tags"
        ],
        "summary": "Update Tag",
        "description": "Update a tag's name or other information.",
        "operationId": "updateTag",
        "security": [
          {
            "bearerAuth": [
              "tags:write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Tag ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpdateRequest"
              }
            },
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag updated successfully",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowedError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Use API key as Bearer token.\n\nFormat: `Authorization: Bearer YOUR_API_KEY`\n\nAPI keys can be generated from the Inblog dashboard.\n"
      }
    },
...
    "schemas": {
      "JsonApiVersion": {
        "type": "object",
        "properties": {
          "jsonapi": {
            "type": "object",
            "properties": {
              "version": {
                "type": "string",
                "default": "1.0",
                "example": "1.0"
              }
            },
            "required": [
              "version"
            ]
          }
        },
        "required": [
          "jsonapi"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "HTTP status code"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "title": {
            "type": "string",
            "description": "Error title"
          },
          "detail": {
            "type": "string",
            "description": "Detailed error message"
          }
        },
        "required": [
          "status",
          "code",
          "title"
        ]
      },
      "ErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonApiVersion"
          },
          {
            "type": "object",
            "properties": {
              "errors": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "required": [
              "errors"
            ]
          }
        ]
      },
...
      "TagResource": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Resource type identifier\n- Fixed value according to JSON:API standard\n- Always \"tags\"\n",
            "enum": [
              "tags"
            ],
            "example": "tags"
          },
          "id": {
            "type": "string",
            "description": "Tag ID"
          },
          "attributes": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "Tag name (case-sensitive, must be unique)",
                "maxLength": 50,
                "example": "JavaScript"
              },
              "slug": {
                "type": "string",
                "description": "URL-friendly unique identifier\n- Only lowercase letters, numbers, and hyphens allowed\n- Can be automatically generated from name\n",
                "pattern": "^[a-z0-9-]+$",
                "example": "javascript"
              }
            }
          }
        }
      },
      "TagListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonApiVersion"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TagResource"
                }
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "TagResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonApiVersion"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/TagResource"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "TagCreateRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonApiVersion"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "tags"
                    ]
                  },
                  "attributes": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "description": "Tag name (required)\n- Maximum 50 characters\n- Must be unique\n",
                        "maxLength": 50,
                        "example": "React"
                      },
                      "slug": {
                        "type": "string",
                        "description": "URL slug (required)\n- Only lowercase letters, numbers, and hyphens allowed\n- Must be unique\n",
                        "pattern": "^[a-z0-9-]+$",
                        "example": "react"
                      }
                    },
                    "required": [
                      "name",
                      "slug"
                    ]
                  }
                },
                "required": [
                  "type",
                  "attributes"
                ]
              }
            }
          }
        ]
      },
      "TagUpdateRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonApiVersion"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "tags"
                    ]
                  },
                  "attributes": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "slug": {
                        "type": "string"
                      }
                    }
                  }
                },
                "required": [
                  "type"
                ]
              }
            }
          }
        ]
      },
...
    },
    "responses": {
...
      "UnauthorizedError": {
        "description": "Authentication failed",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "jsonapi": {
                "version": "1.0"
              },
              "errors": [
                {
                  "status": "401",
                  "code": "unauthorized",
                  "title": "Unauthorized",
                  "detail": "Missing or invalid Authorization header. Expected 'Authorization: Bearer <API_KEY>'"
                }
              ]
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Insufficient permissions",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "jsonapi": {
                "version": "1.0"
              },
              "errors": [
                {
                  "status": "403",
                  "code": "insufficient_permissions",
                  "title": "Insufficient Permissions",
                  "detail": "This operation requires posts:write permission"
                }
              ]
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "jsonapi": {
                "version": "1.0"
              },
              "errors": [
                {
                  "status": "404",
                  "code": "not_found",
                  "title": "Not Found",
                  "detail": "The requested resource could not be found"
                }
              ]
            }
          }
        }
      },
      "MethodNotAllowedError": {
        "description": "Method not allowed",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "jsonapi": {
                "version": "1.0"
              },
              "errors": [
                {
                  "status": "405",
                  "code": "method_not_allowed",
                  "title": "Method Not Allowed",
                  "detail": "The requested method is not allowed for this endpoint"
                }
              ]
            }
          }
        }
      },
      "BadRequestError": {
        "description": "Bad request",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "jsonapi": {
                "version": "1.0"
              },
              "errors": [
                {
                  "status": "400",
                  "code": "validation_error",
                  "title": "Validation Error",
                  "detail": "Missing required fields: title, slug"
                }
              ]
            }
          }
        }
      },
      "ConflictError": {
        "description": "Resource conflict",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "jsonapi": {
                "version": "1.0"
              },
              "errors": [
                {
                  "status": "409",
                  "code": "conflict",
                  "title": "Conflict",
                  "detail": "A resource with this identifier already exists"
                }
              ]
            }
          }
        }
      },
      "InternalServerError": {
        "description": "서버 내부 오류",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "jsonapi": {
                "version": "1.0"
              },
              "errors": [
                {
                  "status": "500",
                  "code": "internal_server_error",
                  "title": "Internal Server Error",
                  "detail": "An unexpected error occurred"
                }
              ]
            }
          }
        }
      }
    }
  }
}